+++ app/src/main/java/com/zotto/kds/database/m_table/Data.kt
@@ -67,7 +67,7 @@ data class Data(
     val rem_om_id: Int,
     val remote_product_id: String,
     val restaurant_group_id: Int,
-    val restaurant_id: Int,
+    val restaurant_id: Long,
     val rts_flag: Int,
     val scaling_qty: Int,
     val sku: String,
